FILO fixes * fix mixup issues between grub_printf and printf * try to unify driver interface * drop a couple of duplicate endianess functions * fix serial port init failure * fix creation of device names from grub interface device names (somewhat) * fix timer calculation and overflow issues * Clean up Makefile and build.sh script
git-svn-id: svn://svn.coreboot.org/filo/trunk/filo@57 205573c7-8e05-0410-a9f4-b951d841fc81

+++ main/elfload.c
@@ -135,7 +135,7 @@ static int load_segments(Elf_phdr *phdr, int phnum,
 	unsigned long checksum_offset)
 {
     unsigned long bytes;
-    unsigned int start_time, time;
+    u64 start_time, time;
     int i;
 
     bytes = 0;
